+++ src/main/java/net/browsertime/tool/BrowserConfig.java
@@ -2,5 +2,6 @@ package net.browsertime.tool;
 
 public enum BrowserConfig {
   windowSize, // <width>x<height>
-  userAgent // user agent string
+  userAgent, // user agent string
+  proxyHost // proxy server url
 }