+++ src/main/java/net/browsertime/tool/datacollector/UserTimingDataCollector.java
@@ -44,7 +44,7 @@ public class UserTimingDataCollector extends TimingDataCollector {
   private static final String LIST_PAGE_DEFINED_MEASUREMENTS =
       "return window.performance.getEntriesByType('measure');";
 
-  private boolean shouldAddMeasurementsForUserMarks;
+  private final boolean shouldAddMeasurementsForUserMarks;
 
   public UserTimingDataCollector(boolean shouldAddMeasurementsForUserMarks) {
     this.shouldAddMeasurementsForUserMarks = shouldAddMeasurementsForUserMarks;
@@ -66,8 +66,9 @@ public class UserTimingDataCollector extends TimingDataCollector {
 
     if (marks != null) {
       for (Map mark : marks) {
-        String name = (String) mark.get("name");
-        double startTime = (Double) mark.get("startTime");
+        MapAdapter ma = new MapAdapter(mark);
+        String name = ma.asString("name");
+        double startTime = ma.asDouble("startTime");
         results.addMark(new TimingMark(name, startTime));
       }
     }
@@ -86,8 +87,9 @@ public class UserTimingDataCollector extends TimingDataCollector {
 
       if (marks != null) {
         for (Map mark : marks) {
-          String name = (String) mark.get("name");
-          double startTime = (Double) mark.get("startTime");
+          MapAdapter ma = new MapAdapter(mark);
+          String name = ma.asString("name");
+          double startTime = ma.asDouble("startTime");
           results.addMeasurement(new TimingMeasurement(name, 0, startTime));
         }
       }
@@ -97,9 +99,10 @@ public class UserTimingDataCollector extends TimingDataCollector {
 
     if (measurements != null) {
       for (Map measurement : measurements) {
-        String name = (String) measurement.get("name");
-        double startTime = (Double) measurement.get("startTime");
-        double duration = (Double) measurement.get("duration");
+        MapAdapter ma = new MapAdapter(measurement);
+        String name = ma.asString("name");
+        double startTime = ma.asDouble("startTime");
+        double duration = ma.asDouble("duration");
         results.addMeasurement(new TimingMeasurement(name, startTime, duration));
       }
     }