+++ src/main/java/net/browsertime/tool/guice/ChromeModule.java
@@ -24,13 +24,11 @@
 package net.browsertime.tool.guice;
 
 
-import com.google.inject.Provider;
 import net.browsertime.tool.BrowserConfig;
 import net.browsertime.tool.datacollector.ChromeDataCollector;
 import net.browsertime.tool.datacollector.TimingDataCollector;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
+import net.browsertime.tool.webdriver.ChromeDriverProvider;
+import net.browsertime.tool.webdriver.WebDriverProvider;
 
 import java.util.Map;
 
@@ -38,7 +36,6 @@ import java.util.Map;
  * Setup a module that uses Chrome.
  */
 public class ChromeModule extends AbstractBrowserModule {
-
   public ChromeModule(Map<BrowserConfig, String> browserConfiguration) {
     super(browserConfiguration);
   }
@@ -46,34 +43,7 @@ public class ChromeModule extends AbstractBrowserModule {
   @Override
   protected void configure() {
     super.configure();
-    bind(WebDriver.class).toProvider(DRIVER_PROVIDER);
+    bind(WebDriverProvider.class).toInstance(new ChromeDriverProvider(browserConfiguration));
     bind(TimingDataCollector.class).to(ChromeDataCollector.class);
   }
-
-  private final Provider<WebDriver> DRIVER_PROVIDER = new Provider<WebDriver>() {
-    @Override
-    public WebDriver get() {
-      return new ChromeDriver(createChromeOptions());
-    }
-
-    private ChromeOptions createChromeOptions() {
-      ChromeOptions options = new ChromeOptions();
-
-      // see http://peter.sh/experiments/chromium-command-line-switches/
-      String config = browserConfiguration.get(BrowserConfig.userAgent);
-      if (config != null) {
-        options.addArguments("--user-agent" + "=" + config);
-      }
-
-      config = browserConfiguration.get(BrowserConfig.windowSize);
-      if (config != null) {
-        config = config.replace("x", ",");
-        options.addArguments("--window-size" + "=" + config);
-      }
-
-      options.addArguments("--window-position=0,0");
-
-      return options;
-    }
-  };
 }