+++ src/main/java/net/browsertime/tool/run/Main.java
@@ -28,6 +28,7 @@ import com.google.inject.AbstractModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Module;
+import net.browsertime.tool.BrowserTimeException;
 import net.browsertime.tool.guice.ChromeModule;
 import net.browsertime.tool.guice.FireFoxModule;
 import net.browsertime.tool.guice.InternetExplorerModule;
@@ -36,8 +37,8 @@ import net.browsertime.tool.guice.XMLResultModule;
 import net.browsertime.tool.serializer.Serializer;
 import net.browsertime.tool.serializer.SerializerFactory;
 import net.browsertime.tool.timingrunner.TimingRunner;
-import net.browsertime.tool.timingrunner.TimingRunnerException;
 import net.browsertime.tool.timings.TimingSession;
+import net.browsertime.tool.webdriver.WebDriverValidationException;
 import org.apache.commons.cli.ParseException;
 
 import java.io.IOException;
@@ -83,9 +84,13 @@ public class Main {
     } catch (IOException e) {
       commandStatus = ERROR;
       printSyntaxError("Error creating output file: " + e.getMessage());
-    } catch (TimingRunnerException e) {
+    } catch (WebDriverValidationException e) {
       commandStatus = ERROR;
-      printSyntaxError("Error occurred while performing timing: " + e.getMessage());
+      printSyntaxError(e.getMessage());
+    } catch (BrowserTimeException e) {
+      commandStatus = ERROR;
+      // This is something serious, print the stacktrace so we know what is happening
+      e.printStackTrace();
     }
 
     if (shouldShowUsage) {
@@ -95,7 +100,7 @@ public class Main {
     return commandStatus;
   }
 
-  private void run(URL url, TimingConfig config) throws IOException {
+  private void run(URL url, TimingConfig config) throws IOException, BrowserTimeException {
     Injector injector = createInjector(config);
 
     TimingRunner timingRunner = injector.getInstance(TimingRunner.class);